Show panel nav watchers item conditionally
