+++ frontend/app/components/routing/wp-details/wp-details.controller.js
@@ -127,14 +127,7 @@ function WorkPackageDetailsController($scope, $state, workPackage, I18n, RELATIO
     }
   }
 
-  //TODO: Remove
-  $scope.canViewWorkPackageWatchers = function() {
-    return !!($scope.workPackage && $scope.workPackage.embedded.watchers !== undefined);
-  };
-
-
   // toggles
-
   $scope.toggleStates = {
     hideFullDescription: true,
     hideAllAttributes: true
